refactor name lognavbar
